+++ finding-gate/cases/design/src/Anonymous/NocHost.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace Corpus\Design\Anonymous;
+
+class NocHost
+{
+    public function build(): object
+    {
+        return new class extends NocParent {
+        };
+    }
+}